+++ Interface/JSY8869/main.py
@@ -1,7 +1,14 @@
+import sys
+
+from PyQt5 import QtWidgets
+
 import main_page
 
 def main():
-    main_page.PyShine_THREADS_APP()
+    app = QtWidgets.QApplication(sys.argv)
+    mainWindow = main_page.PyShine_THREADS_APP()
+    mainWindow.show()
+    sys.exit(app.exec_())
 
 if __name__ == '__main__':
     main()
