+++ Interface/JSY8869/main_page.py
@@ -1,15 +1,12 @@
-from PyQt5 import QtCore, QtWidgets
-import sys
-import time
-
 from PyQt5.QtGui import QPalette, QBrush, QImage
 from PyQt5.QtWidgets import *
 from PyQt5 import uic
 
 import weather
+from ThreadClass import ThreadClass
 
 thread_ui = uic.loadUiType("ui/threads.ui")[0]
-class PyShine_THREADS_APP(QMainWindow,QWidget, thread_ui):
+class PyShine_THREADS_APP(QMainWindow, QWidget, thread_ui):
     def __init__(self):
         self.thread = {}
         super(PyShine_THREADS_APP,self).__init__()
@@ -47,13 +44,13 @@ class PyShine_THREADS_APP(QMainWindow,QWidget, thread_ui):
     def start_worker_1(self):
         self.thread[1] = ThreadClass(parent=None, index=1)
         self.thread[1].start()
-        self.thread[1].any_signal.connect(self.my_function)
+        self.thread[1].any_signal.connect(self.progress_bar_count)
         self.pushButton.setEnabled(False)
 
     def start_worker_2(self):
         self.thread[2] = ThreadClass(parent=None, index=2)
         self.thread[2].start()
-        self.thread[2].any_signal.connect(self.my_function)
+        self.thread[2].any_signal.connect(self.progress_bar_count)
         self.pushButton_3.setEnabled(False)
 
     def stop_worker_1(self):
@@ -89,7 +86,7 @@ class PyShine_THREADS_APP(QMainWindow,QWidget, thread_ui):
         except:
             self.textEdit.setText("날씨 조회 실패 (인터넷 상태나 주소 설정을 확인해주세요.)")
 
-    def my_function(self, counter):
+    def progress_bar_count(self, counter):
 
         cnt = counter
         index = self.sender().index
@@ -101,31 +98,3 @@ class PyShine_THREADS_APP(QMainWindow,QWidget, thread_ui):
             self.progressBar_2.setValue(cnt)
             if cnt == 100:
                 self.pushButton_2.setEnabled(True)
-
-
-
-class ThreadClass(QtCore.QThread):
-    any_signal = QtCore.pyqtSignal(int)
-
-    def __init__(self, parent=None, index=0):
-        super(ThreadClass, self).__init__(parent)
-        self.index = index
-        self.is_running = True
-
-    def run(self):
-        cnt = 0
-        while (True):
-            cnt += 1
-            if cnt == 101:
-                return
-            time.sleep(0.01)
-            self.any_signal.emit(cnt)
-    def stop(self):
-        self.any_signal.emit(0)
-        self.is_running = False
-        self.terminate()
-
-app = QtWidgets.QApplication(sys.argv)
-mainWindow = PyShine_THREADS_APP()
-mainWindow.show()
-sys.exit(app.exec_())
